% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{load_hugmycoa}
\alias{load_hugmycoa}
\title{Load phyloseq object from HuGMycoA}
\usage{
load_hugmycoa(level = c("Genus", "SH"), filterdepth = TRUE)
}
\arguments{
\item{level}{Character. "Genus" or "SH"}

\item{filterdepth}{Logical. TRUE = filter depth, FALSE = no filter}
}
\value{
phyloseq object
}
\description{
Load phyloseq object from HuGMycoA
}
